# Proposed fix to add setters tests