--- a/include/ada/encoding_type.h
+++ b/include/ada/encoding_type.h
@@ -9,11 +9,13 @@ namespace ada {
    *
    * @see https://encoding.spec.whatwg.org/#encodings
    */
-  enum encoding_type {
+  enum class encoding_type {
     UTF8,
     UTF_16LE,
     UTF_16BE,
   };
+
+  ada_warn_unused std::string to_string(encoding_type type);
 
 } // ada namespace