--- a/include/ada/parser.h
+++ b/include/ada/parser.h
@@ -17,7 +17,7 @@ namespace ada::parser {
 
   url parse_url(std::string user_input,
                 std::optional<ada::url> base_url = std::nullopt,
-                ada::encoding_type encoding = UTF8,
+                ada::encoding_type encoding = ada::encoding_type::UTF8,
                 std::optional<ada::url> optional_url = std::nullopt,
                 std::optional<ada::state> state_override = std::nullopt);