--- a/include/ada/state.h
+++ b/include/ada/state.h
@@ -3,7 +3,7 @@
 
 namespace ada {
 
-  enum state {
+  enum class state {
     AUTHORITY,
     SCHEME_START,
     SCHEME,
@@ -25,6 +25,8 @@ namespace ada {
     OPAQUE_PATH,
     PORT,
   };
+
+  ada_warn_unused std::string to_string(ada::state s);
 
 } // ada namespace