--- a/src/implementation.cpp
+++ b/src/implementation.cpp
@@ -14,18 +14,12 @@ namespace ada {
     }
   }
 
-  // TODO: This should take std::string_view input or at the very least const std::string& input.
-  ada_warn_unused url parse(std::string input,
+  ada_warn_unused url parse(std::string_view input,
                             std::optional<ada::url> base_url,
                             ada::encoding_type encoding,
                             std::optional<ada::state> state) noexcept {
-
-    return ada::parser::parse_url(std::move(input), std::move(base_url), encoding, std::nullopt, state);
-  }
-
-  // TODO: This is worse than useless.
-  ada_warn_unused url parse(std::string_view input) noexcept {
-    return ada::parser::parse_url(std::string(input));
+    // TODO std::move(base_url) might be unwise. Check.
+    return ada::parser::parse_url(input, std::move(base_url), encoding, std::nullopt, state);
   }
 
   /*