--- a/tests/wpt_tests.cpp
+++ b/tests/wpt_tests.cpp
@@ -138,44 +138,45 @@ bool setters_tests_encoding() {
 
       if (category == "protocol") {
         std::string_view expected = element["expected"]["protocol"];
-        ada::set_scheme(base, std::string(new_value));
+        ada::set_scheme(base, std::string{new_value});
         TEST_ASSERT(base.scheme + ":", expected, "Protocol");
       }
       else if (category == "username") {
         std::string_view expected = element["expected"]["username"];
-        ada::set_username(base, std::string(new_value));
+        ada::set_username(base, std::string{new_value});
         TEST_ASSERT(base.username, expected, "Username");
       }
       else if (category == "password") {
         std::string_view expected = element["expected"]["password"];
-        ada::set_password(base, std::string(new_value));
+        ada::set_password(base, std::string{new_value});
         TEST_ASSERT(base.password, expected, "Password");
       }
+//TODO: Why is this commented out?
 //      else if (category == "host") {
 //        std::string_view expected = element["expected"]["host"];
-//        ada::set_host(base, std::string(new_value));
+//        ada::set_host(base, std::string{new_value});
 //        TEST_ASSERT(base.host.value_or(ada::url_host{ada::BASIC_DOMAIN, ""}).entry, expected, "Host");
 //      }
       else if (category == "port") {
         std::string_view expected = element["expected"]["port"];
-        ada::set_port(base, std::string(new_value));
+        ada::set_port(base, std::string{new_value});
         auto normalized = base.port.has_value() ? std::to_string(*base.port) : "";
         TEST_ASSERT(normalized, expected, "Port");
       }
       else if (category == "pathname") {
         std::string_view expected = element["expected"]["pathname"];
-        ada::set_pathname(base, std::string(new_value));
+        ada::set_pathname(base, std::string{new_value});
         TEST_ASSERT(base.path, expected, "Path");
       }
       else if (category == "search") {
         std::string_view expected = element["expected"]["search"];
-        ada::set_search(base, std::string(new_value));
+        ada::set_search(base, std::string{new_value});
         auto normalized = !base.query.value_or("").empty() ? "?" + base.query.value() : "";
         TEST_ASSERT(normalized, expected, "Search");
       }
       else if (category == "hash") {
         std::string_view expected = element["expected"]["hash"];
-        ada::set_hash(base, std::string(new_value));
+        ada::set_hash(base, std::string{new_value});
         auto normalized = !base.fragment.value_or("").empty() ? "#" + *base.fragment : "";
         TEST_ASSERT(normalized, expected, "Fragment");
       }
@@ -239,12 +240,12 @@ bool urltestdata_encoding() {
       ada::url base_url;
       if (!object["base"].get(base)) {
         std::cout << "base=" << base << std::endl;
-        base_url = ada::parse(std::string{base});
+        base_url = ada::parse(base);
       }
       bool failure = false;
       ada::url input_url = (!object["base"].get(base)) ?
-      ada::parse(std::string{input}, ada::parse(std::string{base}))
-      : ada::parse(std::string{input});
+      ada::parse(input, ada::parse(base))
+      : ada::parse(input);
 
       if (!object["failure"].get(failure)) {
         TEST_ASSERT(input_url.is_valid, !failure, "Failure");