This is a PR on add-setters-tests. It improves the performance somewhat, but we still have a large regression compared to the main branch. Prior to merging any of this into the main branch, we should understand the cause of the large regression.

```
This PR:
BasicBench_AdaURL       7985 ns         7985 ns        83977 time/byte=9.88189ns time/url=725.87ns url/s=1.37766M/s
BasicBench_AdaURL       7934 ns         7933 ns        87971 time/byte=9.81829ns time/url=721.198ns url/s=1.38658M/s
BasicBench_AdaURL       7904 ns         7903 ns        88095 time/byte=9.7807ns time/url=718.437ns url/s=1.39191M/s
BasicBench_AdaURL       7947 ns         7946 ns        88526 time/byte=9.83446ns time/url=722.386ns url/s=1.3843M/s

the original setters PR
BasicBench_AdaURL       8432 ns         8432 ns        83380 time/byte=10.4356ns time/url=766.543ns url/s=1.30456M/s
BasicBench_AdaURL       8425 ns         8424 ns        82965 time/byte=10.4263ns time/url=765.856ns url/s=1.30573M/s
BasicBench_AdaURL       8433 ns         8431 ns        83155 time/byte=10.435ns time/url=766.498ns url/s=1.30464M/s
BasicBench_AdaURL       8383 ns         8382 ns        83157 time/byte=10.3741ns time/url=762.023ns url/s=1.3123M/s

Main branch:
BasicBench_AdaURL       7242 ns         7241 ns        96974 time/byte=8.96132ns time/url=658.25ns url/s=1.51918M/s
BasicBench_AdaURL       7195 ns         7194 ns        97116 time/byte=8.90391ns time/url=654.033ns url/s=1.52898M/s
BasicBench_AdaURL       7220 ns         7218 ns        97100 time/byte=8.93345ns time/url=656.203ns url/s=1.52392M/s
BasicBench_AdaURL       7268 ns         7267 ns        96063 time/byte=8.99431ns time/url=660.673ns url/s=1.51361M/s
```