@lemire 2023-01-13
Merging directly to [add-setters-tests](https://github.com/ada-url/ada/tree/add-setters-tests) where I will be doing more work.